pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "vindicara"
version = "0.1.0"
description = "Runtime security for autonomous AI. The control plane for AI agents in production."
readme = "README.md"
license = "Apache-2.0"
requires-python = ">=3.11"
authors = [{ name = "Vindicara", email = "eng@vindicara.io" }]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Security",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Typing :: Typed",
]
dependencies = [
    "httpx>=0.27.0,<1.0",
    "pydantic>=2.7.0,<3.0",
    "pydantic-settings>=2.3.0,<3.0",
    "structlog>=24.1.0,<26.0",
]

[project.optional-dependencies]
api = [
    "fastapi>=0.115.0,<1.0",
    "mangum>=0.19.0,<1.0",
    "uvicorn>=0.30.0,<1.0",
    "boto3>=1.35.0,<2.0",
    "boto3-stubs[dynamodb,s3,events]>=1.35.0,<2.0",
]
cdk = [
    "aws-cdk-lib>=2.150.0,<3.0",
    "constructs>=10.0.0,<11.0",
]
dev = [
    "pytest>=8.3.0,<9.0",
    "pytest-asyncio>=0.24.0,<1.0",
    "pytest-cov>=5.0.0,<6.0",
    "hypothesis>=6.100.0,<7.0",
    "ruff>=0.6.0,<1.0",
    "mypy>=1.11.0,<2.0",
    "pip-audit>=2.7.0,<3.0",
]

[project.urls]
Homepage = "https://vindicara.io"
Repository = "https://github.com/vindicara/vindicara"
Documentation = "https://docs.vindicara.io"

[tool.hatch.build.targets.wheel]
packages = ["src/vindicara"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
asyncio_mode = "auto"
markers = ["adversarial: tests for adversarial/bypass inputs"]

[tool.ruff]
target-version = "py311"
line-length = 100
src = ["src", "tests"]

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "S", "B", "A", "C4", "PT", "SIM", "TCH", "RUF"]
ignore = [
    "S101",   # assert in tests
    "B008",   # Depends() in function defaults is standard FastAPI pattern
    "A002",   # shadowing 'input' builtin is intentional SDK API design
    "N818",   # exception names without Error suffix are intentional public API
]

[tool.ruff.lint.isort]
known-first-party = ["vindicara"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true
plugins = ["pydantic.mypy"]

[tool.coverage.run]
source = ["vindicara"]

[tool.coverage.report]
fail_under = 80
show_missing = true
